requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
numpy>=1.24.0
scikit-learn>=1.3.0
pydantic>=2.0.0
pyyaml>=6.0
openpyxl>=3.1.0
